ability to change user roles and protect delete functionality

+++ convex/files.ts
@@ -34,7 +34,8 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
   }
 
   const hasAccess =
-    user.orgId.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some((item) => item.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId);
 
   if (!hasAccess) {
     return null;
@@ -118,6 +119,15 @@ export const deleteFile = mutation({
     if (!access) {
       throw new ConvexError("no access to file");
     }
+
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === "admin";
+
+    if (!isAdmin) {
+      throw new ConvexError("no access to file");
+    }
+
     await ctx.db.delete(args.fileId);
   },
 });
